Extract out non-control specific tests into new files
